Move calibrator-flats BDD tests behind bdd-infra rp-harness feature
Introduces an optional `rp-harness` feature on the `bdd-infra` workspace
crate that exposes OmniSim, rp config builder, webhook receiver, test
orchestrator, and an rmcp MCP test client. rp's own BDD tests migrate to
use this feature and calibrator-flats now owns its end-to-end BDD suite
under `services/calibrator-flats/tests/`, spawning rp via
`rp_harness::start_rp(sibling_service_dir(...), ...)`.

The feature gate keeps the four lightweight BDD consumers (filemonitor,
qhy-focuser, ppba-driver, sentinel) off axum/reqwest/rmcp — verified via
`cargo tree`. `cargo-rail` now re-runs calibrator-flats BDD only when
rp, calibrator-flats, or bdd-infra changes, so future workflow plugins
will not bloat rp's test run time.

Also shook out a latent test-isolation bug: `RpConfigBuilder::build()`
now PID+seq-suffixes `data_directory` and `session_state_file` so two
BDD binaries running in the same `cargo test` invocation cannot clobber
each other's session state.

Documentation (calibrator-flats.md, rp.md, testing.md) updated to
describe the per-plugin BDD convention and the `rp-harness` feature.

322 BDD scenarios pass; unit tests and clippy green.
