+++ services/calibrator-flats/Cargo.toml
@@ -27,6 +27,13 @@ axum = { workspace = true }
 rmcp = { workspace = true, features = ["client", "transport-streamable-http-client-reqwest"] }
 
 [dev-dependencies]
+bdd-infra = { workspace = true, features = ["rp-harness"] }
+cucumber = { workspace = true }
+libc = { workspace = true }
 tempfile = { workspace = true }
 tokio-test = { workspace = true }
 cargo-husky = { workspace = true }
+
+[[test]]
+name = "bdd"
+harness = false
