+++ services/calibrator-flats/tests/bdd.rs
@@ -1,4 +1,7 @@
-//! BDD test entry point for rp service
+//! BDD test entry point for the calibrator-flats service.
+//!
+//! These tests spawn three processes — OmniSim, rp, and calibrator-flats —
+//! and drive the flat calibration workflow end-to-end via rp's REST API.
 
 #[path = "bdd/world.rs"]
 mod world;
@@ -8,12 +11,15 @@ mod steps;
 
 bdd_infra::bdd_main! {
     use cucumber::World as _;
-    use world::RpWorld;
+    use world::CalibratorFlatsWorld;
 
-    RpWorld::cucumber()
+    CalibratorFlatsWorld::cucumber()
         .after(|_feature, _rule, _scenario, _finished, maybe_world| {
             Box::pin(async move {
                 if let Some(world) = maybe_world {
+                    if let Some(cf) = world.calibrator_flats.as_mut() {
+                        cf.stop().await;
+                    }
                     if let Some(rp) = world.rp.as_mut() {
                         rp.stop().await;
                     }
