+++ services/calibrator-flats/tests/bdd/steps/mod.rs
@@ -1,0 +1,3 @@
+//! BDD step definitions for the calibrator-flats service.
+
+pub mod flat_calibration_steps;
