+++ services/rp/tests/bdd/steps/equipment_steps.rs
@@ -2,8 +2,10 @@
 
 use cucumber::{given, then, when};
 
-use crate::steps::infrastructure::{OmniSimHandle, ServiceHandle};
-use crate::world::{CameraConfig, FilterWheelConfig, RpWorld};
+use bdd_infra::rp_harness::{CameraConfig, FilterWheelConfig, OmniSimHandle};
+use bdd_infra::ServiceHandle;
+
+use crate::world::RpWorld;
 
 // --- Given steps ---
