+++ services/rp/tests/bdd/steps/mod.rs
@@ -6,10 +6,7 @@ pub mod camera_info_steps;
 pub mod cover_calibrator_steps;
 pub mod equipment_steps;
 pub mod event_steps;
-pub mod flat_calibration_steps;
 pub mod image_stats_steps;
-pub mod infrastructure;
-pub mod mcp_test_client;
 pub mod session_steps;
 pub mod tls_steps;
 pub mod tool_steps;
